requirements.txt
python-dotenv==1.0.1
requests==2.32.3
black==24.8.0
playwright==1.48.0
